[Runtime] bug fix for StringTable grow
Summary:
For StringTable, it expects that the Node shouldn't be changed after it
is inserted to a hash table.
If the string object in the node is changed via Reflection or Unsafe, it
breaks the assumption and JVM crashes when the string table grows.
To avoid the crash, we drop the 'dead ' Node when the table grows.

Use -XX:+IgnoreInvalidEntryAtStringTableExpansion to enable the feature.

Test Plan:
test/hotspot/jtreg/ajdk/TestStringTableGrow.java

Reviewers: lei.yul, sanhong.lsh, chuansheng.lcs, kuaiwei.kw, yunyao.zxl

Issue: https://aone.alibaba-inc.com/task/25294062

CR: https://code.aone.alibaba-inc.com/xcode/jdk11/codereview/2509172